+++ services/autoscaling/src/simcore_service_autoscaling/core/application.py
@@ -10,8 +10,8 @@ from .._meta import (
     APP_STARTED_BANNER_MSG,
 )
 from ..api.routes import setup_api_routes
+from ..modules.auto_scaling_task import setup as setup_background_task
 from ..modules.docker import setup as setup_docker
-from ..modules.dynamic_scaling import setup as setup_background_task
 from ..modules.ec2 import setup as setup_ec2
 from ..modules.rabbitmq import setup as setup_rabbitmq
 from ..modules.redis import setup as setup_redis
