+++ services/autoscaling/src/simcore_service_autoscaling/modules/dynamic_scaling_core.py
@@ -7,20 +7,11 @@ from datetime import datetime, timedelta, timezone
 from typing import cast
 
 from fastapi import FastAPI
-from models_library.generated_models.docker_rest_api import (
-    Availability,
-    Node,
-    NodeState,
-    Task,
-)
+from models_library.generated_models.docker_rest_api import Task
 from pydantic import parse_obj_as
 from types_aiobotocore_ec2.literals import InstanceTypeType
 
-from ..core.errors import (
-    Ec2InstanceNotFoundError,
-    Ec2InvalidDnsNameError,
-    Ec2TooManyInstancesError,
-)
+from ..core.errors import Ec2InstanceNotFoundError, Ec2TooManyInstancesError
 from ..core.settings import ApplicationSettings, get_application_settings
 from ..models import (
     AssociatedInstance,
@@ -31,18 +22,12 @@ from ..models import (
 )
 from ..utils import ec2, utils_docker
 from ..utils.dynamic_scaling import (
-    associate_ec2_instances_with_nodes,
     ec2_startup_script,
-    node_host_name_from_ec2_private_dns,
     try_assigning_task_to_instances,
     try_assigning_task_to_node,
     try_assigning_task_to_pending_instances,
 )
-from ..utils.rabbitmq import (
-    log_tasks_message,
-    post_autoscaling_status_message,
-    progress_tasks_message,
-)
+from ..utils.rabbitmq import log_tasks_message, progress_tasks_message
 from .docker import get_docker_client
 from .ec2 import get_ec2_client
 
@@ -408,112 +393,9 @@ async def _scale_up_cluster(
     return cluster
 
 
-async def _try_attach_pending_ec2s(app: FastAPI, cluster: Cluster) -> Cluster:
-    """label the drained instances that connected to the swarm which are missing the monitoring labels"""
-    new_found_instances: list[AssociatedInstance] = []
-    still_pending_ec2s: list[EC2InstanceData] = []
-    app_settings = get_application_settings(app)
-    assert app_settings.AUTOSCALING_EC2_INSTANCES  # nosec
-    for instance_data in cluster.pending_ec2s:
-        try:
-            node_host_name = node_host_name_from_ec2_private_dns(instance_data)
-            if new_node := await utils_docker.find_node_with_name(
-                get_docker_client(app), node_host_name
-            ):
-                # it is attached, let's label it, but keep it as drained
-                new_node = await utils_docker.tag_node(
-                    get_docker_client(app),
-                    new_node,
-                    tags=utils_docker.get_docker_tags(app_settings),
-                    available=False,
-                )
-                new_found_instances.append(AssociatedInstance(new_node, instance_data))
-            else:
-                still_pending_ec2s.append(instance_data)
-        except Ec2InvalidDnsNameError:
-            logger.exception("Unexpected EC2 private dns")
-    # NOTE: first provision the reserve drained nodes if possible
-    all_drained_nodes = (
-        cluster.drained_nodes + cluster.reserve_drained_nodes + new_found_instances
-    )
-    return dataclasses.replace(
-        cluster,
-        drained_nodes=all_drained_nodes[
-            app_settings.AUTOSCALING_EC2_INSTANCES.EC2_INSTANCES_MACHINES_BUFFER :
-        ],
-        reserve_drained_nodes=all_drained_nodes[
-            : app_settings.AUTOSCALING_EC2_INSTANCES.EC2_INSTANCES_MACHINES_BUFFER
-        ],
-        pending_ec2s=still_pending_ec2s,
-    )
-
-
-async def _analyze_current_cluster(app: FastAPI) -> Cluster:
-    app_settings = get_application_settings(app)
-    assert app_settings.AUTOSCALING_NODES_MONITORING  # nosec
-    assert app_settings.AUTOSCALING_EC2_INSTANCES  # nosec
-
-    # get current docker nodes (these are associated (active or drained) or disconnected)
-    docker_nodes: list[Node] = await utils_docker.get_monitored_nodes(
-        get_docker_client(app),
-        node_labels=app_settings.AUTOSCALING_NODES_MONITORING.NODES_MONITORING_NODE_LABELS,
-    )
-
-    # get the EC2 instances we have
-    existing_ec2_instances = await get_ec2_client(app).get_instances(
-        app_settings.AUTOSCALING_EC2_INSTANCES,
-        ec2.get_ec2_tags(app_settings),
-    )
-
-    terminated_ec2_instances = await get_ec2_client(app).get_instances(
-        app_settings.AUTOSCALING_EC2_INSTANCES,
-        ec2.get_ec2_tags(app_settings),
-        state_names=["terminated"],
-    )
-
-    attached_ec2s, pending_ec2s = await associate_ec2_instances_with_nodes(
-        docker_nodes, existing_ec2_instances
-    )
-
-    def _is_node_up_and_available(node: Node, availability: Availability) -> bool:
-        assert node.Status  # nosec
-        assert node.Spec  # nosec
-        return bool(
-            node.Status.State == NodeState.ready
-            and node.Spec.Availability == availability
-        )
-
-    def _node_not_ready(node: Node) -> bool:
-        assert node.Status  # nosec
-        return bool(node.Status.State != NodeState.ready)
-
-    all_drained_nodes = [
-        i
-        for i in attached_ec2s
-        if _is_node_up_and_available(i.node, Availability.drain)
-    ]
-
-    cluster = Cluster(
-        active_nodes=[
-            i
-            for i in attached_ec2s
-            if _is_node_up_and_available(i.node, Availability.active)
-        ],
-        drained_nodes=all_drained_nodes[
-            app_settings.AUTOSCALING_EC2_INSTANCES.EC2_INSTANCES_MACHINES_BUFFER :
-        ],
-        reserve_drained_nodes=all_drained_nodes[
-            : app_settings.AUTOSCALING_EC2_INSTANCES.EC2_INSTANCES_MACHINES_BUFFER
-        ],
-        pending_ec2s=pending_ec2s,
-        terminated_instances=terminated_ec2_instances,
-        disconnected_nodes=[n for n in docker_nodes if _node_not_ready(n)],
-    )
-    logger.info("current state: %s", f"{cluster=}")
-    return cluster
-
-
-async def _scale_cluster(app: FastAPI, cluster: Cluster) -> Cluster:
+async def scale_cluster_with_labelled_services(
+    app: FastAPI, cluster: Cluster
+) -> Cluster:
     app_settings = get_application_settings(app)
     assert app_settings.AUTOSCALING_NODES_MONITORING  # nosec
     assert app_settings.AUTOSCALING_EC2_INSTANCES  # nosec
@@ -539,23 +421,3 @@ async def _scale_cluster(app: FastAPI, cluster: Cluster) -> Cluster:
         cluster = await _try_scale_down_cluster(app, cluster)
 
     return cluster
-
-
-async def _cleanup_disconnected_nodes(app: FastAPI, cluster: Cluster) -> Cluster:
-    await utils_docker.remove_nodes(get_docker_client(app), cluster.disconnected_nodes)
-    return dataclasses.replace(cluster, disconnected_nodes=[])
-
-
-async def cluster_scaling_from_labelled_services(app: FastAPI) -> None:
-    """Check that there are no pending tasks requiring additional resources in the cluster (docker swarm)
-    If there are such tasks, this method will allocate new machines in AWS to cope with
-    the additional load.
-    """
-
-    cluster = await _analyze_current_cluster(app)
-    cluster = await _cleanup_disconnected_nodes(app, cluster)
-    cluster = await _try_attach_pending_ec2s(app, cluster)
-    cluster = await _scale_cluster(app, cluster)
-
-    # inform on rabbit about status
-    await post_autoscaling_status_message(app, cluster)
